+++ component/todoList.js
@@ -1,6 +1,31 @@
-import { filterState, handleCount, renderActiveItems, renderCompleteItems } from "./todoCount.js";
-import { todoListEl, saveToDos, toDos, filterToDos } from "../todoApp.js";
-export const handleTodoItemClick=(event)=>{
+import { todoListEl, toDoInput } from "./constant.js";
+import {toDos, filterToDos} from "../init.js";
+import { saveToDos } from "./todoLocalStorage.js";
+import { handleCount } from "./todoCount.js";
+import { renderFromFilter } from "./todoFilter.js";
+
+const addToDos = (item)=>{
+    todoListEl.insertAdjacentHTML("beforeend",renderTodoItemTemplate(item));
+    toDoInput.value="";
+}
+
+const addToItems = (item)=>{
+    toDos.push(item);
+}
+
+const renderTodoItemTemplate=(item)=>{
+    return (
+     `<li data-id="${item.id}" class="${item.completed ? "completed" : ""}">
+        <div class="view">
+            <input class="toggle" type="checkbox" ${item.completed ? "checked" : ""}/>
+            <label class="label">${item.title}</label>
+            <button class="destroy"></button>
+        </div>
+        <input class="edit" value="${item.title}" />
+    </li>`);
+};
+
+const handleTodoItemClick=(event)=>{
     const targetClass = event.target.className.split(" ");
 
     if(targetClass[0] === "toggle") handleComplete(event);
@@ -8,10 +33,32 @@ export const handleTodoItemClick=(event)=>{
 
 }
 
+const handleNewTodoSubmit = async (event)=>{
+    const newItem = {
+        id: Date.now(),
+        title: event.target.value,
+        completed: false
+    }
+
+    addToItems(newItem);
+    addToDos(newItem);
+    handleCount(toDos.length);
+    saveToDos();
+    renderFromFilter();
+    /*
+    if(filterState === "completed"){
+        renderCompleteItems();
+    } else if (filterState === "active"){
+       renderActiveItems();
+    }*/
+
+};
+
 const handleComplete=(event)=>{
     event.target.closest("li").classList.toggle("completed");
     event.target.closest("input").classList.toggle("checked");
-    itemsUpdate(event);
+    todoCompleted(event);
+    renderFromFilter();
 }
 
 const handleDestory=(event)=>{
@@ -19,14 +66,14 @@ const handleDestory=(event)=>{
     const li = event.target.parentNode.parentNode;
     removeFromItems(li);
     todoListEl.removeChild(li);
-  
     handleCount(toDos.length);
 }
 
+
 const removeFromItems=(li)=>{
     try{
-        const testItemId = li.dataset.id;
-        filterToDos(toDos,testItemId);
+        const todoItemId = li.dataset.id;
+        filterToDos(toDos,todoItemId);
  
         
     }catch(error){
@@ -35,11 +82,12 @@ const removeFromItems=(li)=>{
     
 }
 
-const itemsUpdate=(event)=>{
+const todoCompleted=(event)=>{
 
     const currentLi = event.target.closest('li');
-
+    
     const currentItemId = currentLi.dataset.id;
+
     for(let obj of toDos){
         if(obj.completed === false && obj.id === parseInt(currentItemId)){
             obj.completed = true;
@@ -48,12 +96,11 @@ const itemsUpdate=(event)=>{
             obj.completed = false;
         }
     }
-    if(filterState === "completed"){
-        renderCompleteItems();
-    } else if (filterState === "active"){
-        renderActiveItems();
-    }
+
     saveToDos();
 
 }
 
+
+
+export {addToDos, addToItems, handleTodoItemClick, handleNewTodoSubmit};
