+++ index.html
@@ -35,6 +35,6 @@
       </main>
     </div>
 
-    <script type="module" src="init.js"></script>
+    <script type="module" src="./init.js"></script>
   </body>
 </html>
